EventType.jump but there's a weird bug.
Sometimes, the jump event doesn't function properly and the respective node containing the jump event will be triggered twice as a result.

It's got to do with the NextNode() callbacks. Needs further inspection. This is gonna be a hard one and I seriously hope it doesn't become a wall I run into.