"EventType.jump" bug fixed.
Added a condition where the player will fully stop after the distance between it and the current node is less than 0.1 units.

TO DO: enact the "willStop" boolean when the distance between the player and the node is less than 25% of the total length between the current and the last node instead of always having a fixed number to everything.