Crouching works but conflicts with other events.
Yeah... crouching works but it causes weird issues when coupling "duck" and "rise" events with a "walk" event. (Should I first keep the "walk" event above the "rise" or "duck" events?)

My assumption is that when rising up, the player is triggering the node's respective collider and therefore, the NextNode() callback.

Progress is really slow. But it's better to have something rather than nothing at all.