Code for player start at Node#000.
It works for now but there has to be a much better manner to implement it. On the bright side, I can now spawn the player as a prefab on Start and not necessarily always have a player object present in the level by default.

This also means that I have to change the SampleScene nodes.