Merged camera alignment to a single function.
The "look" event looked really messy. Having a dedicated button each just to align the camera to the path of the nodes and then having another button just to position the camera to the node. Nope.

Removed all of that alignment code from 'Drawer_NodeListItem' and placed it in the RepositionCameraToNode() function of 'Inspector_Node'. It's better to have one button to do everything rather than have more than one button to do each separate thing.