Lots of things I forgot to commit.
Lots of progress with "demo_movement". Added a bunch of prefabs for demo construction props for convenience and building the node path for the player to follow.

Had to tweak the code here and there and let's face it, it's not perfect. Needs a lot of perfection and refactoring after this first build.