Work on "demo_movement" and Node.cs
Some more work on the "demo_movement" scene and added an IF check in Node.cs that prevents the player from colliding and parsing events of nodes it isn't supposed to interact with.

Now I don't have to worry about restricting the node path based on whether the player will interact with them or not.