And the level loading bug still fucking exists.
Fuck everything. This isn't noticed in "demo_movement" but it's almost frequent in "SampleScene". And somehow, it seemed to work pretty fine when checking it out via the Editor until I made a build to test it.

Somehow, faster loading times prevent the restarted level to work as intended. Or maybe it could have something to do with the fact that I am always creating a player right around the "Start" of the GameManager.

I genuinely don't know what to do at this point. I am so goddamn fucking close and this shit has to happen.