"Slowing down" code disabled.
Trying to come up with a solution to write a decent equation to slow down the player nicely irrespective of frame rates was just taking really long and I was growing frustrated. Fucking Thursdays, miserable weather and lack of sleep... and I really wanted to push a build of this. Had to take the painful decision to remove the code responsible for slowing down the player before entering a node with a "stop" event.

It looks so odd seeing the player abruptly stop and move. Just fuck it. I'll fix it in the next build.